BREAKING changed syntax of admFocus and admBlur
A previous build had broken admFocus and admBlur, but on reimplementing
realised they weren't working correctly in the first place - they were
being called every time the cursor flashed. (In fact the code calling
them didn't work at all and they were being called incidentally due to a
bug.)

Testing revealed that there was no way to implement them using the
current syntax (that is, passing a line of code for execution) - the
& binding, which is used to allow execution in the parent scope, can't
seem to access grandparent scope under nested directives each binding
with &. As such, needed to switch to = binding, and so could only
transfer objects, not code for execution.